单元测试,数据库限制

+++ app/Command/Tools/PhpunitCommand.php
@@ -12,7 +12,6 @@ namespace App\Command\Tools;
 
 use Hyperf\Command\Annotation\Command;
 use Hyperf\Command\Command as HyperfCommand;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
 /**
@@ -25,8 +24,8 @@ class PhpunitCommand extends HyperfCommand
     public function configure()
     {
         parent::configure();
-        $this->addOption('filter', 'f',InputOption::VALUE_OPTIONAL, '测试类型,单个指定,默认测试全部', 'all');
-        $this->addOption('reset', 'r',InputOption::VALUE_OPTIONAL, '重建数据库,且跑完整的测试流程', 'no');
+        $this->addOption('filter', 'f', InputOption::VALUE_OPTIONAL, '测试类型,单个指定,默认测试全部', 'all');
+        $this->addOption('reset', 'r', InputOption::VALUE_OPTIONAL, '重建数据库,且跑完整的测试流程', 'no');
         $this->setDescription('单元测试');
     }
 
@@ -34,20 +33,24 @@ class PhpunitCommand extends HyperfCommand
     {
         $filter = $this->input->getOption('filter');
         $reset = $this->input->getOption('reset');
-        if (!in_array($reset, ['yes', 'no'])) {
+        if (! in_array($reset, ['yes', 'no'])) {
             $this->output->error('--reset yes/no');
             return;
         }
         $testExec = 'vendor/bin/co-phpunit -c phpunit.xml --colors=always';
         if ($reset == 'yes') {
+            if (stripos(env('DB_DATABASE'), '_test') === false) {
+                $this->output->error('请使用测试数据库(xx_test)做单元测试');
+                return;
+            }
             exec('php bin/hyperf.php migrate:fresh --seed');
-            exec("rm -rf runtime/container");
+            exec('rm -rf runtime/container');
             exec($testExec, $output);
         } else {
             if ($filter != 'all') {
                 $testExec = "{$testExec} '--filter={$filter}'";
             }
-            exec("rm -rf runtime/container");
+            exec('rm -rf runtime/container');
             exec($testExec, $output);
         }
         foreach ($output as $row) {
